requirements.txt
django-cors-headers>=4.3,<5.0
Django>=5.x.x
djangorestframework>=3.15.x
python-dotenv>=1.x.x
psycopg2-binary>=2.9,<3.0
celery
redis
yfinance
flower
djangorestframework-simplejwt